lte/alt1250: Change usrsock interface with nuttx update
The usrsock interface is changed with nuttx update.

cbe9352f41 net/usrsock: remove the connections limit
Replace xid type to uint64_t to avoid limitation of connections.

451c53daa4 usrsock: Move event field to usrsock_message_common_s
Change the structure which event field is defined.

+++ lte/alt1250/usock_handlers/alt1250_getsocknamehdlr.c
@@ -45,7 +45,7 @@ static int postproc_getsockname(FAR struct alt1250_s *dev,
                                 FAR struct alt_container_s *reply,
                                 FAR struct usock_s *usock,
                                 FAR int32_t *usock_result,
-                                FAR uint8_t *usock_xid,
+                                FAR uint64_t *usock_xid,
                                 FAR struct usock_ackinfo_s *ackinfo,
                                 unsigned long arg)
 {
@@ -126,7 +126,7 @@ int nextstep_getsockname(FAR struct alt1250_s *dev,
                          FAR struct alt_container_s *reply,
                          FAR struct usock_s *usock,
                          FAR int32_t *usock_result,
-                         FAR uint8_t *usock_xid,
+                         FAR uint64_t *usock_xid,
                          FAR struct usock_ackinfo_s *ackinfo,
                          unsigned long arg)
 {
@@ -153,7 +153,7 @@ int nextstep_getsockname(FAR struct alt1250_s *dev,
 int usockreq_getsockname(FAR struct alt1250_s *dev,
                          FAR struct usrsock_request_buff_s *req,
                          FAR int32_t *usock_result,
-                         FAR uint8_t *usock_xid,
+                         FAR uint64_t *usock_xid,
                          FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_getsockname_s *request = &req->request.name_req;
