+++ lte/alt1250/usock_handlers/alt1250_ioctl_normal.c
@@ -64,7 +64,7 @@ static int postproc_reportnet(FAR struct alt1250_s *dev,
                               FAR struct alt_container_s *reply,
                               FAR struct usock_s *usock,
                               FAR int32_t *usock_result,
-                              FAR uint8_t *usock_xid,
+                              FAR uint64_t *usock_xid,
                               FAR struct usock_ackinfo_s *ackinfo,
                               unsigned long is_async)
 {
@@ -86,7 +86,7 @@ static int postproc_radioon(FAR struct alt1250_s *dev,
                             FAR struct alt_container_s *reply,
                             FAR struct usock_s *usock,
                             FAR int32_t *usock_result,
-                            FAR uint8_t *usock_xid,
+                            FAR uint64_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo,
                             unsigned long arg)
 {
@@ -117,7 +117,7 @@ static int postproc_actpdn(FAR struct alt1250_s *dev,
                               FAR struct alt_container_s *reply,
                               FAR struct usock_s *usock,
                               FAR int32_t *usock_result,
-                              FAR uint8_t *usock_xid,
+                              FAR uint64_t *usock_xid,
                               FAR struct usock_ackinfo_s *ackinfo,
                               unsigned long is_async)
 {
@@ -150,7 +150,7 @@ int postproc_fwgetversion(FAR struct alt1250_s *dev,
                           FAR struct alt_container_s *reply,
                           FAR struct usock_s *usock,
                           FAR int32_t *usock_result,
-                          FAR uint8_t *usock_xid,
+                          FAR uint64_t *usock_xid,
                           FAR struct usock_ackinfo_s *ackinfo,
                           unsigned long arg)
 {
@@ -203,7 +203,7 @@ int send_lapi_command(FAR struct alt1250_s *dev,
 int usockreq_ioctl_normal(FAR struct alt1250_s *dev,
                               FAR struct usrsock_request_buff_s *req,
                               FAR int32_t *usock_result,
-                              FAR uint8_t *usock_xid,
+                              FAR uint64_t *usock_xid,
                               FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_ioctl_s *request = &req->request.ioctl_req;
